DEMOS-1538: added file in the middle to fetch data
For deliverable/id route
