+++ client/src/pages/DemonstrationDetail/index.tsx
@@ -1,1 +1,2 @@
 export { DemonstrationDetail } from "./DemonstrationDetail";
+export { DeliverableDetailRoute } from "./DeliverableDetailRoute";
